test: ckBTC transaction service
