--- frontend/src/tests/lib/services/icrc-transactions.services.spec.ts
@@ -1,1 +1,0 @@
-export {};
