+++ frontend/src/tests/mocks/icrc-transactions.mock.ts
@@ -1,5 +1,5 @@
-import type { IcrcTransactionWithId } from "@dfinity/ledger";
-import type { Principal } from "@dfinity/principal";
+import type { IcrcTransaction, IcrcTransactionWithId } from "@dfinity/ledger";
+import { Principal } from "@dfinity/principal";
 
 export interface IcrcCandidAccount {
   owner: Principal;
@@ -28,3 +28,30 @@ export const createIcrcTransactionWithId = (
     ],
   },
 });
+
+const fakeAccount = {
+  owner: Principal.fromText("aaaaa-aa"),
+  subaccount: [] as [],
+};
+
+const mockIcrcTransaction: IcrcTransaction = {
+  kind: "transfer",
+  timestamp: BigInt(12354),
+  burn: [],
+  mint: [],
+  transfer: [
+    {
+      to: fakeAccount,
+      from: fakeAccount,
+      memo: [],
+      created_at_time: [BigInt(123)],
+      amount: BigInt(33),
+      fee: [BigInt(1)],
+    },
+  ],
+};
+
+export const mockIcrcTransactionWithId: IcrcTransactionWithId = {
+  id: BigInt(123),
+  transaction: mockIcrcTransaction,
+};
